+++ testing/test_full.py
@@ -872,3 +872,45 @@ def test_augmentation_functions():
         output_tensor = global_augs_dict[aug]()(input_tensor)
         assert output_tensor != None, "Augmentation should work"
     print("passed")
+
+
+def test_checkpointing_segmentation_rad_2d(device):
+    print("Starting 2D Rad segmentation tests for metrics")
+    # read and parse csv
+    parameters = parseConfig(
+        testingDir + "/config_segmentation.yaml", version_check=False
+    )
+    training_data, parameters["headers"] = parseTrainingCSV(
+        inputDir + "/train_2d_rad_segmentation.csv"
+    )
+    parameters = populate_header_in_parameters(parameters, parameters["headers"])
+    parameters["patch_size"] = patch_size["2D"]
+    parameters["num_epochs"] = 1
+    parameters["nested_training"]["testing"] = 1
+    parameters["model"]["dimension"] = 2
+    parameters["model"]["class_list"] = [0, 255]
+    parameters["model"]["amp"] = True
+    parameters["model"]["num_channels"] = 3
+    parameters["metrics"] = ["dice", "hausdorff", "hausdorff95"]
+    parameters["model"]["architecture"] = "unet"
+    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    TrainingManager(
+        dataframe=training_data,
+        outputDir=outputDir,
+        parameters=parameters,
+        device=device,
+        reset_prev=True,
+    )
+    parameters["num_epochs"] = 2
+    parameters["nested_training"]["validation"] = -2
+    parameters["nested_training"]["testing"] = 1
+    TrainingManager(
+        dataframe=training_data,
+        outputDir=outputDir,
+        parameters=parameters,
+        device=device,
+        reset_prev=False,
+    )
+    shutil.rmtree(outputDir)  # overwrite previous results
+
+    print("passed")